Add files via upload
Considering the difference between male voice and female voice, this network only used female audio data to build the training set.

In this version, I built the validation set to spare overfitting during training. While the size of the validation set is too small(80) to fit the lstm network. The validation set did not cover all the phonemes (46). It only had 45 phonemes. 

The problem here is that there are not enough data. As is estimated before, about 200 training examples can cover all the phonemes. Considering the training set, validation set and the test set to be built, the overall 460 examples are not enough to train this network. 

There are 46 phonemes in all. Some of them are quite rare in daily sentences. To collect enough data is a critical problem for this project.