Add files via upload
To spare time of building data set before training the network, I created this file to build the data set for this project. This file is for the female data only. All the audio data are converted into one cell array. The size of the cell array(XFemale) is 460*1, each cell included one audio example. All the label files are converted into one cell array(YFemale) . The size of this cell array is also 460*1. Each cell included one example.